requirements.txt
aiosqlite
httpx
pyrogram
python-dotenv
telethon
